First batch of Unit Tests
Adds around 60 unit test cases to cover a good proportion of Uplift.
The .bat script is a basic shortcut file to run the tests in one line, it
should be translatable to shell easily enough.
NOTE: the tests can also be ran from VisualStudio if the NUnit Adapter has
been installed.