Package Nuking test suite
Add a package nuking test suite to the integration tests.

NOTE: The Nuking functionnality was tested through the LocalHandler
directly and neither through the CLI nor through the Unity interface. The
latter is not testable outside of Unity, and the passing arguments to the
CLI through code is not what we want to do.