Move MD5 storing to UpliftManager
After the installation, the lockfile MD5 changed from what was
previously tracked, causing another installation to be triggered, which
obviously was without effect, but still caused some debugging and some
actions that should not be fired. Consequently, the MD5 storing has been
moved to the UpliftManager so it can register the new state after
installing, and not cause a "double install".