Menu Items: Make installation from menu clearer
A lot of feedback tend to show that there was a confusion regarding what
was doing what between "Debug/Install from lockfile" and "Install
Dependencies". This moves the lockfile installation out of the debug
section as it is not a debug-only feature and is used quite frequently.
Moreover, this clarifies the messages to better explain the difference
between the installation methods.